Balls to it: drop authentication requirement entirely
I've been going backwards and forwards over this decision for a while
now, but I think I'm settled. Windows authentication for SMB is totally
unpredictable.

The basic premise seems to be that the local and remote user accounts
must share the same username and the user must be granted access on the
server and on the share on the client. This is seems to leave us with
weird permissions on written files and gives us a whole host of
authentication failures within IIS which just don't make sense.

We'll go with host-based authentication and call it a day until someone
with more patience than myself volunteers to fix it.